- started new implementation for UWP+DX11

+++ src/rendersystems/DirectX/11/default-shaders/PixelShader_TexturedAlphaMap.hlsl
@@ -1,8 +1,8 @@
-/// @version 4.1
+/// @version 5.2
 
 #include "PixelShader_Textured.hlsli"
 
-min16float4 main(PixelShaderInput input) : SV_Target
+float4 main(PixelShaderInput input) : SV_Target
 {
-	return min16float4(input.color.rgb, input.color.a * cTexture.Sample(cSampler, input.tex).r);
+	return float4(input.color.rgb, input.color.a * cTexture.Sample(cSampler, input.tex).r);
 }
