+++ src/rendersystems/DirectX/11/default-shaders/VertexShader_Plain.hlsl
@@ -1,17 +1,17 @@
-/// @version 4.1
+/// @version 5.2
 
 #include "VertexShader.hlsli"
 #include "PixelShader_Plain.hlsli"
 
 struct VertexShaderInput
 {
-	min16float3 position : POSITION;
+	float3 position : POSITION;
 };
 
 PixelShaderInput main(VertexShaderInput input)
 {
 	PixelShaderInput vertexShaderOutput;
-	vertexShaderOutput.position = mul(min16float4(input.position, (min16float)1.0), cMatrix);
+	vertexShaderOutput.position = mul(float4(input.position, 1.0), cMatrix);
 	vertexShaderOutput.color = cSystemColor;
 	vertexShaderOutput.lerpAlpha = cLerpAlpha;
 	return vertexShaderOutput;
